requirements.txt
aiogram==3.12.0
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.29.0
python-dotenv==1.0.1
apscheduler==3.10.4